stocks/requirements.txt
Flask==2.0.1
Werkzeug==2.2.2
requests==2.26.0
python-dotenv==1.0.1
pymongo==4.3.3
